minor IPFIX cleanups
