+++ src/gui/types/conf.rs
@@ -4,8 +4,8 @@ use crate::gui::types::config_window::ConfigWindow;
 use crate::gui::types::export_pcap::ExportPcap;
 use crate::gui::types::favorite::Favorites;
 use crate::gui::types::filters::Filters;
+use crate::gui::types::ipfix_socket::MyIpfixSocket;
 use crate::gui::types::settings::Settings;
-use crate::networking::ipfix::MyIpfixSocket;
 use crate::networking::types::capture_context::CaptureSourcePicklist;
 use crate::networking::types::config_device::ConfigDevice;
 use crate::networking::types::data_representation::DataRepr;
