+++ src/gui/types/mod.rs
@@ -3,6 +3,7 @@ pub mod config_window;
 pub mod export_pcap;
 pub mod favorite;
 pub mod filters;
+pub mod ipfix_socket;
 pub mod message;
 pub mod settings;
 pub mod timing_events;
